+++ app/src/main/java/com/echoist/linkedout/App.kt
@@ -2,16 +2,16 @@ package com.echoist.linkedout
 
 import android.app.Application
 import android.content.Context
-import android.util.Log
 import androidx.room.Room
+import com.echoist.linkedout.api.AuthInterceptor
 import com.echoist.linkedout.api.BookMarkApi
+import com.echoist.linkedout.api.ErrorHandlingInterceptor
 import com.echoist.linkedout.api.EssayApi
 import com.echoist.linkedout.api.SignUpApi
 import com.echoist.linkedout.api.SocialSignUpApi
 import com.echoist.linkedout.api.StoryApi
 import com.echoist.linkedout.api.SupportApi
 import com.echoist.linkedout.api.UserApi
-import com.echoist.linkedout.page.myLog.Token
 import com.echoist.linkedout.room.EssayStorageDB
 import com.echoist.linkedout.room.EssayStoreDao
 import com.squareup.moshi.Moshi
@@ -22,9 +22,7 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.HiltAndroidApp
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Response
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import javax.inject.Singleton
@@ -32,37 +30,6 @@ import javax.inject.Singleton
 @HiltAndroidApp
 class App : Application() {
     // 애플리케이션 클래스 안에서 모듈을 정의하고 설치합니다.
-}
-class ErrorHandlingInterceptor : Interceptor {
-    override fun intercept(chain: Interceptor.Chain): Response {
-        val request = chain.request()
-        val response = chain.proceed(request)
-
-        if (!response.isSuccessful) {
-            // 여기서 response.code()나 response.body() 등을 활용해 에러 핸들링 로직을 작성합니다.
-            when (response.code) {
-                401 -> {
-                    // 인증 오류 처리 (예: 토큰 만료 시 재로그인 유도)
-                    Log.e("intercept err", "intercept: ${response.code}")
-                    if (response.headers["x-access-token"].isNullOrEmpty()) { //x-access-token이 안들어오는경우 재로그인필요.
-                        AuthManager.isReAuthenticationRequired.value = true
-                    } else {
-                        Log.e("intercept err", "intercept: ${response.headers["x-access-token"]!!}")
-                        Token.accessToken = response.headers["x-access-token"]!!
-                    }
-                }
-                500 -> {
-                    // 서버 오류 처리
-                    Log.e("intercept err", "intercept: ${response.code}")
-                }
-                else -> {
-                    // 기타 에러 처리
-                    Log.e("intercept err", "intercept: ${response.code}")
-                }
-            }
-        }
-        return response
-    }
 }
 
 @Module
@@ -80,7 +47,8 @@ object AppModule {
 //
         // OkHttpClient에 인터셉터를 추가
         val httpClient = OkHttpClient.Builder()
-            .addInterceptor(ErrorHandlingInterceptor())
+            .addInterceptor(AuthInterceptor()) //전역 헤더 인터셉터 추가.
+            .addInterceptor(ErrorHandlingInterceptor()) // 전역 에러핸들링 인터셉터 추가
             .build()
 
         return Retrofit.Builder()
